refactor(app): restructure flutter project & modularize widgets
